Move legacy UDP Server callback to dcs-service
